Refactor command menu
